requirements.txt
numpy>=1.23
pandas>=1.5
scipy>=1.9
scikit-learn>=1.2
pillow>=9.0
matplotlib>=3.6
torch
torchvision
torchsummary
git+https://github.com/openai/CLIP.git

